chatbot/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai>=0.7.2
duckduckgo-search>=6.2.7
pytz
pydantic-settings